fix(widgets): rename BurgerMenu.tsx
